+++ src/enums/TinyMCESource.php
@@ -7,15 +7,15 @@
 namespace spicyweb\tinymce\enums;
 
 /**
- * Pseudo-enum for representing the source to load TinyMCE from.
+ * Enum for representing the source to load TinyMCE from.
  *
  * @package spicyweb\tinymce\enums
  * @author Spicy Web <plugins@spicyweb.com.au>
- * @since 1.4.0
+ * @since 2.0.0
  */
-abstract class TinyMCESource
+enum TinyMCESource: string
 {
-    public const Default = 'default';
-    public const TinyCloud = 'tinyCloud';
-    public const Custom = 'custom';
+    case Default = 'default';
+    case TinyCloud = 'tinyCloud';
+    case Custom = 'custom';
 }